pyproject.toml
[project]
name = "axion-viewer"
version = "1.0.0"
description = "API Flask para conversão de documentos Markdown em relatórios HTML e PDF formatados"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Marcos", email = "marcos@example.com" }]

dependencies = [
    "Flask>=3.0.0",
    "Flask-CORS>=4.0.0",
    "markdown>=3.5.1",
    "beautifulsoup4>=4.12.2",
    "Jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "Werkzeug>=3.0.1",
    "WeasyPrint>=62.3",
    "Pillow>=10.1.0",
    "cffi>=1.16.0",
    "pydyf>=0.11.0",
    "gunicorn>=21.2.0",
    "supabase>=2.0.0",
    "PyJWT[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.8",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils"]
include = ["app.py", "config.py"]

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "node_modules",
    "frontend",
    "data",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "frontend/*", "data/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
